\docType{data}
\name{auto}
\alias{auto}
\title{auto dataset from stata}
\format{A dataframe with 74 observations on the following 12 variables

\describe{

\item{"make"}{maker of car.}

\item{"price"}{price of car.}

\item{"mpg"}{miles per gallon.}

\item{"rep78"}{.}

\item{"headroom"}{headroom.}

\item{"trunk"}{trunk.}

\item{"weight"}{weight.}

\item{"length"}{length.}

\item{"turn"}{turn.}

\item{"displacement"}{displacement.}

\item{"gear_ratio"}{gear ratio.}

\item{"foreign"}{price of car.}

}}
\description{
  auto dataset from stata
}
\keyword{datasets}

