\docType{package}
\name{grpack-package}
\alias{grpack-package}
\title{grpack}
\description{
  mmest is a package that contains useful functions,
  classes, and methods for econometric analysis.
}
\details{
  \tabular{ll}{ Package: \tab grpack\cr Type: \tab
  Package\cr Version: \tab 0.1-1\cr Date: \tab
  2011-12-01\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Giuseppe Ragusa \email{gragusa@luiss.it},

  Maintainer: Giuseppe Ragusa \email{gragusa@luiss.it}
}
\keyword{package}

