\docType{methods}
\name{modelInfo}
\alias{modelInfo}
\alias{modelInfo.summary.coxph}
\alias{modelInfo.summary.gee}
\alias{modelInfo.summary.glm}
\alias{modelInfo.summary.lm}
\alias{modelInfo.summary.lrm}
\alias{modelInfo.summary.reg}
\alias{modelInfo.summary.tobit}
\title{summary methods used by \code{"outreg2"}.}
\usage{
  modelInfo.summary.reg(x)

  modelInfo.summary.lm(x)

  modelInfo.summary.glm(x)

  modelInfo.summary.tobit(x)

  modelInfo.summary.gee(x)

  modelInfo.summary.coxph(x)

  modelInfo.summary.lrm(x)
}
\value{
  A \code{model.info} object.
}
\description{
  summary methods used by \code{"outreg2"}.
}
\seealso{
  \code{\link{outreg2}} and \code{\link{apsrtable}}.
}

